@@ -1,11 +1,17 @@
 module synyx.de/tales
 
-go 1.14
+go 1.22
 
 require (
 	github.com/gorilla/mux v1.8.1
 	github.com/mozillazg/go-slugify v0.2.0
-	github.com/mozillazg/go-unidecode v0.1.1 // indirect
-	github.com/stretchr/testify v1.8.4
-	olympos.io/encoding/edn v0.0.0-20200308123125-93e3b8dd0e24
+	github.com/stretchr/testify v1.9.0
+	olympos.io/encoding/edn v0.0.0-20201019073823-d3554ca0b0a3
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/mozillazg/go-unidecode v0.2.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
