@@ -3,13 +3,13 @@ package web
 import (
 	"bytes"
 	"encoding/json"
-	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
 	"os"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+
 	"synyx.de/tales/pkg/project"
 )
 
@@ -21,7 +21,7 @@ type TestClient struct {
 }
 
 func NewTestClient(t *testing.T) *TestClient {
-	dir, err := ioutil.TempDir("", "tales-test")
+	dir, err := os.MkdirTemp("", "tales-test")
 	assert.NoError(t, err)
 	handler := NewServer(dir, "")
 	repo := &project.FilesystemRepository{
